Minor refactor for SQLFederationEngine and SQLFederationDataTypeBuilder
